requirements.txt
Pygments==2.1.3
python-magic==0.4.12
isc-dhcp-leases==0.8.1
requests==2.4.3
PyYAML==3.11
Flask==0.11.1
Flask-Script==2.0.5
celery==3.1.23
SQLAlchemy==1.0.13
marshmallow-sqlalchemy==0.8.1
tornado==4.4.1
Werkzeug==0.11.11
six==1.10.0
singledispatch==3.4.0.3
Flask-Markdown==0.3
Flask-SQLAlchemy==2.1
flask-marshmallow==0.7.0
Sphinx==1.4.6
semver==2.7.3
pexpect==4.0.1
paramiko==1.17.2
aeon-venos>=0.5.1
retrying